worked on folder structure
