+++ src/components/Header/Header.js
@@ -1,14 +1,11 @@
-import MenuIcon from "@material-ui/icons/Menu";
-import YouTubeIcon from "@material-ui/icons/YouTube";
-import MicIcon from "@material-ui/icons/Mic";
 import VideocamIcon from "@material-ui/icons/Videocam";
 import NotificationsNoneIcon from "@material-ui/icons/NotificationsNone";
 import SearchIcon from "@material-ui/icons/Search";
-import React, {useState} from "react";
+import React from "react";
 import "./Header.css";
-import {useAuth} from "../context/auth-context";
+import {useAuth} from "../index";
 import {Link} from "react-router-dom";
-import {useVideo} from "../context/video-context";
+import {useVideo} from "../index";
 export const Header = ({setShowSidebar}) => {
   const {login} = useAuth();
   const {dispatch} = useVideo();
@@ -26,10 +23,6 @@ export const Header = ({setShowSidebar}) => {
       }}
     >
       <div className="nav-left  lg pd-half">
-        {/* <span className="mg-left-half flex row align-items-center ">
-          <YouTubeIcon fontSize={"large"} /> CrunchTube
-        </span> */}
-
         <label className="flex row search-label">
           <input
             className="search-input"
@@ -57,11 +50,6 @@ export const Header = ({setShowSidebar}) => {
             <button style={{color: "white"}} className="avatar-sm">
               {localStorage.getItem("username")?.slice(0, 1).toUpperCase()}
             </button>
-            {/* <img
-              src="https://cdn.discordapp.com/avatars/778699572787675136/bfd4785c0eb8a2be4eae7cc2fec27d34.png?size=128"
-              className="avatar-modified-sm"
-              alt="avatar"
-            /> */}
           </div>
         )}
         {!login && (
