+++ src/components/LikedList/index.js
@@ -1,0 +1,1 @@
+export {LikedList} from "./LikedList";
