+++ src/components/LoginAndSignup/Signup.js
@@ -1,6 +1,6 @@
 import React, {useState, useRef, useEffect} from "react";
 import {Link, useLocation, useNavigate} from "react-router-dom";
-import {useAuth} from "../context/auth-context";
+import {useAuth} from "../index";
 import axios from "axios";
 
 export const Signup = () => {
