+++ src/components/LoginAndSignup/index.js
@@ -1,0 +1,2 @@
+export {Login} from "./Login";
+export {Signup} from "./Signup";
