+++ src/components/PlayLists/index.js
@@ -1,0 +1,2 @@
+export {PlayLists} from "./PlayLists";
+export {PlayListVideoPage} from "./PlayListVideoPage";
