+++ src/components/VideoList/index.js
@@ -1,0 +1,2 @@
+export {VideoList} from "./VideoList";
+export {VideoCard} from "./VideoCard";
