+++ src/components/VideoPage/LikeButton.js
@@ -28,7 +28,6 @@ export const LikeButton = ({
       console.log(response);
     } catch (error) {
       setError(error.response.data.message);
-
       console.log({error});
     }
   };
