+++ src/components/VideoPage/RemoveLikeButton.js
@@ -28,7 +28,6 @@ export const RemoveLikeButton = ({
       console.log(response);
     } catch (error) {
       setError(error.response.data.message);
-
       console.log({error});
     }
   };
