+++ src/components/WatchedList/index.js
@@ -1,0 +1,1 @@
+export {WatchedList} from "./WatchedList";
