+++ src/context/index.js
@@ -1,0 +1,2 @@
+export {useAuth} from "./auth-context";
+export {useVideo} from "./video-context";
